third_party/WebKit/LayoutTests/scrollbars/custom-scrollbar-not-inherited-by-iframe.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<style>
::-webkit-scrollbar {
    width: 150px;
    height: 150px;
}
iframe {
  width: 300px;
  border: 0;
  outline: 1px solid blue;
}
</style>
<iframe srcdoc="
    <style>
        body { margin: 0; padding: 10px; height: 500px; font: 10pt monospace; }
    </style>
  "></iframe>
<script>
async_test(function(t) {
    onload = t.step_func_done(function() {
        var body = document.querySelector("iframe").contentDocument.body;
        assert_equals(285, body.clientWidth);
    });
});
</script>
